requirements.txt
flask
flask-sqlalchemy
requests
flask-migrate
python-dotenv
flask_apscheduler